core and framework fixed

+++ core/src/main/java/com/utd/indoorairmonitor/domain/PeakFlow.kt
@@ -1,4 +1,5 @@
 package com.utd.indoorairmonitor.domain
 
-class PeakFlow(val peakFlow: Float) {
+//Reading (260 - 270 units, best is 300 units)
+class PeakFlow(val peakFlow: Int) {
 }
